# FanApp/FanApp/OpenAPIs/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "FanClubAPI",
    "version": "1.0"
  },
  "paths": {
...
    "/api/Concerts": {
      "get": {
        "tags": [
          "Concerts"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Concerts"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Concerts"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Concerts"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Concerts"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Concerts"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Concerts"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Concerts"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Concerts"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Concerts"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Concerts"
                }
              }
            }
          }
        }
      }
    },
    "/api/Concerts/{id}": {
      "get": {
        "tags": [
          "Concerts"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Concerts"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Concerts"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Concerts"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Concerts"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Concerts"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Concerts"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Concerts"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Concerts"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Members": {
      "get": {
        "tags": [
          "Members"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Members"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Members"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Members"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Members"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Members"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Members"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Members"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Members"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Members"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Members"
                }
              }
            }
          }
        }
      }
    },
    "/api/Members/{id}": {
      "get": {
        "tags": [
          "Members"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Members"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Members"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Members"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Members"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Members"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Members"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Members"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Members"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/News": {
      "get": {
        "tags": [
          "News"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/News"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/News"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/News"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "News"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/News"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/News"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/News"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/News"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/News"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/News"
                }
              }
            }
          }
        }
      }
    },
    "/api/News/{id}": {
      "get": {
        "tags": [
          "News"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/News"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/News"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/News"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "News"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/News"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/News"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/News"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "News"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Albums": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "releaseDate": {
            "$ref": "#/components/schemas/DateOnly"
          },
          "coverUrl": {
            "type": "string",
            "nullable": true
          },
          "songs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Songs"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Concerts": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "concertDate": {
            "$ref": "#/components/schemas/DateOnly"
          },
          "city": {
            "type": "string",
            "nullable": true
          },
          "venue": {
            "type": "string",
            "nullable": true
          },
          "ticketUrl": {
            "type": "string",
            "nullable": true
          },
          "setlists": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Setlists"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "DateOnly": {
        "type": "object",
        "properties": {
          "year": {
            "type": "integer",
            "format": "int32"
          },
          "month": {
            "type": "integer",
            "format": "int32"
          },
          "day": {
            "type": "integer",
            "format": "int32"
          },
          "dayOfWeek": {
            "$ref": "#/components/schemas/DayOfWeek"
          },
          "dayOfYear": {
            "type": "integer",
            "format": "int32",
            "readOnly": true
          },
          "dayNumber": {
            "type": "integer",
            "format": "int32",
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
      "DayOfWeek": {
        "enum": [
          0,
          1,
          2,
          3,
          4,
          5,
          6
        ],
        "type": "integer",
        "format": "int32"
      },
      "Members": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "role": {
            "type": "string",
            "nullable": true
          },
          "joinDate": {
            "$ref": "#/components/schemas/DateOnly"
          },
          "leaveDate": {
            "$ref": "#/components/schemas/DateOnly"
          },
          "bio": {
            "type": "string",
            "nullable": true
          },
          "photoUrl": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "News": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "content": {
            "type": "string",
            "nullable": true
          },
          "publishDate": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
...
      "Setlists": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "concertId": {
            "type": "integer",
            "format": "int32"
          },
          "songId": {
            "type": "integer",
            "format": "int32"
          },
          "songOrder": {
            "type": "integer",
            "format": "int32"
          },
          "concert": {
            "$ref": "#/components/schemas/Concerts"
          },
          "song": {
            "$ref": "#/components/schemas/Songs"
          }
        },
        "additionalProperties": false
      },
      "Songs": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "albumId": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "duration": {
            "type": "string",
            "nullable": true
          },
          "trackNumber": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "lyrics": {
            "type": "string",
            "nullable": true
          },
          "album": {
            "$ref": "#/components/schemas/Albums"
          },
          "setlists": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Setlists"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}